% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Carpool retrieve data from server}
\usage{
get_data()
}
\arguments{
\item{user}{id}
}
\description{
This function allows you to retrieve data from server.
}
\examples{
get_data()
}
\keyword{carpool}

